requirements.txt
langchain>=0.1.0
redis>=5.0.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
PyPDF2>=3.0.0
click>=8.1.0
tiktoken>=0.5.1
flask-cors>=6.0.0
